Fixed project pages: 'bilineareq'
