--- app/projects/biquads/biquads.module.css
@@ -1,26 +1,0 @@
-.container {
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-.content {
-	max-width: 56rem; /* 896px */
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 3.5rem; /* 56px */
-
-	padding-left: 0.75rem; /* 12px */
-  padding-right: 0.75rem; /* 12px */
-	padding-top: 4rem; /* 64px */
-	padding-bottom: 4rem; /* 64px */
-}
-
-@media (min-width: 1024px) {
-	.content {
-	  padding-top: 6rem; /* 96px */
-	  padding-bottom: 6rem; /* 96px */
-	}
-}
