+++ app/projects/biquads/biquads_content.tsx
@@ -1,15 +1,15 @@
 import { MDXRemote } from 'next-mdx-remote/rsc'
 import {
   HRGradient,
-  TextLargeBoldCenter,
   BackToHome
 } from '@/components/layouts'
-import styles from '@/app/projects/biquads/biquads.module.css'
+import styles from '@/app/layout.module.css'
+
+export const dynamic = 'force-dynamic'
 
 export default async function BiquadsContent() {
 
   // MDX text - can be from a local file, database, CMS, fetch, anywhere...
-  // const res = await fetch('http://localhost:3000/markdown/projects/biquads/biquads.mdx')
   const res = await fetch('https://raw.githubusercontent.com/nathanjhood/Biquads/main/README.md')
   const markdown = await res.text()
 
@@ -18,13 +18,9 @@ export default async function BiquadsContent() {
 
       <div className={styles.content}>
 
-        <div className="flex flex-col gap-8 bg-background text-foreground">
+        <div className={styles.flexboxgrid}>
 
-          <HRGradient />
-
-          <MDXRemote source={`[![Biquads](https://github-readme-stats-two-lime-18.vercel.app/api/pin/?username=nathanjhood\&repo=Biquads\&theme=transparent)](https://github.com/StoneyDSP/Biquads)`} />
-
-          <HRGradient />
+          {/* <MDXRemote source={`[![Biquads](https://github-readme-stats-two-lime-18.vercel.app/api/pin/?username=nathanjhood\&repo=Biquads\&theme=transparent)](https://github.com/StoneyDSP/Biquads)`} /> */}
 
           <MDXRemote source={markdown} />
