+++ public/markdown/projects/biquads/biquads.mdx.d.ts
@@ -1,0 +1,3 @@
+export { default } from './biquads.mdx'
+
+export const message: string;
