+++ public/markdown/projects/orfanidisbiquad/orfanidisbiquad.mdx
@@ -1,0 +1,11 @@
+# OrfanidisBiquad
+
+v.1.1.0b.1 (10/06/2022)
+
+[![OrfanidisBiquad](https://github-readme-stats-two-lime-18.vercel.app/api/pin/?username=nathanjhood\&repo=OrfanidisBiquad\&theme=transparent)](https://github.com/StoneyDSP/OrfanidisBiquad)
+
+![Orf-AutoGUI-1-1-0b](https://raw.githubusercontent.com/StoneyDSP/OrfanidisBiquad/master/Res/Orf-AutoGUI-1-1-0b.png)
+
+Available for testing!
+
+A refactor is on the cards to solve the maths involved - the current method does work, but I would not suggest it is an optimal approach, more rather a working prototype.
