+++ public/sitemap.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <sitemap>
+    <loc>https://stoneydsp.com/www_sitemap.xml</loc>
+  </sitemap>
+  <sitemap>
+    <loc>https://stoneydsp.com/sitemap.xml</loc>
+  </sitemap>
+</sitemapindex>
