+++ public/www_sitemap.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url>
+    <loc>https://www.stoneydsp.com</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/about</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/contact</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/privacy-policy</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/terms-of-service</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/audioplugin-svf</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/bilineareq</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/biquads</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/cxxwin</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/msys2-toolchain</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/nonlinearfilters</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/orfanidisbiquad</loc>
+  </url>
+  <url>
+    <loc>https://www.stoneydsp.com/projects/ubento</loc>
+  </url>
+</urlset>
